Some more big changes (hopefully the last for a while).
All of the London-specific code has now been stripped out of the main distribution and stuck into a "mod" - the site now uses "mods" as developed by riseup.net coders to allow some kinds of functionality to be overridden in plugin-like fashion (it's based on Rails Engines, but functions as sort of Engines-in-reverse).  

There is also a configuration menu so that admins can now manipulate site settings in the admin UI. Configuration values are loaded once at application start, and changing the values when the site is running saves the new setting to the database and changes the static Hyperactive.whatever value.  This saves us from hitting the database whenever we want to find out the site name, making things nice and quick.